Move application submission to POST to /cas2/submissions
CAS2 Applications are now submitted using a POST to the submissions endpoint:

`POST /cas2/submissions` with
  `{ translatedDocument: JSON, applicationId: UUID}`

rather than

`POST /cas2/applications/{applicationId}/submission` with
  `{ translatedDocument: JSON, type: 'CAS2'}`

Representing a submitted application as a "submission" resource with a type of `CAS2SubmittedApplication`
allows us to avoid adding to complexity to the `ApplicationsController`.

Note also that:

- the `applicationId` is now received as the `applicationId`
  url param rather than being in the path.

- the `type` param is no longer part of the interface, as this was not being used.
  We aren't using the "discriminator" triggered by
  `type: CAS2` as we have already moved out of the JPA
  inheritance linking all the CAS applications. CAS2 has its
  own `Cas2Application` entity and stand-alone
  `cas_2_applications` db table.
