+++ src/main/resources/static/codegen/built-cas2-api-spec.yml
@@ -130,13 +130,32 @@ paths:
           $ref: '#/components/responses/403Response'
         500:
           $ref: '#/components/responses/500Response'
-  /applications/{applicationId}/submission:
+  /submissions:
+    get:
+      tags:
+        - Operations on submitted CAS2 applications (Assessors)
+      summary: List summaries of all submitted CAS2 applications
+      responses:
+        200:
+          description: successful operation
+          content:
+            'application/json':
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/Cas2SubmittedApplicationSummary'
+        401:
+          $ref: '#/components/responses/401Response'
+        403:
+          $ref: '#/components/responses/403Response'
+        500:
+          $ref: '#/components/responses/500Response'
     post:
       tags:
         - Operations on CAS2 applications
       summary: Submits a CAS2 Application
       parameters:
-        - in: path
+        - in: query
           name: applicationId
           required: true
           description: Id of the application
@@ -159,26 +178,6 @@ paths:
             'application/problem+json':
               schema:
                 $ref: '#/components/schemas/ValidationError'
-        401:
-          $ref: '#/components/responses/401Response'
-        403:
-          $ref: '#/components/responses/403Response'
-        500:
-          $ref: '#/components/responses/500Response'
-  /submissions:
-    get:
-      tags:
-        - Operations on submitted CAS2 applications (Assessors)
-      summary: List summaries of all submitted CAS2 applications
-      responses:
-        200:
-          description: successful operation
-          content:
-            'application/json':
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/Cas2SubmittedApplicationSummary'
         401:
           $ref: '#/components/responses/401Response'
         403:
